gateway/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
circuitbreaker==2.0.0
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
flask-prometheus-metrics==1.0.0
grpcio==1.59.3
grpcio-tools==1.59.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0
protobuf==4.24.2
psycopg2==2.9.9
pymongo==4.6.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
